[REF] default USER and PASSWORD to odoo
Defaulting USER to postgres does not make sense as odoo prevents
creating a database with the user `postgres` (however, there's a bug
that allows it when set by environment variable, but still it does not
work later because Odoo does not use the environment variable to find
the databases in the `list_dbs` method).

Defaulting PASSWORD to odoo makes sense because it's what we set in
the documentationm and is even mandatory since we chose to pass the pg
connections parameters and we did not set a default value for it.